# swagger.json
{"openapi":"3.0.0","paths":{"/info/roomList":{"get":{"operationId":"InfoController_getRoomList","summary":"get room list","parameters":[],"responses":{"200":{"description":"No error."},"401":{"description":"Unauthorized."}},"tags":["info"],"security":[{"token":[]}]}},"/info/newRoom":{"post":{"operationId":"InfoController_newRoom","summary":"Add new meeting room info.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Room"}}}},"responses":{"200":{"description":"No error."},"401":{"description":"Unauthorized"}},"tags":["info"],"security":[{"token":[]}]}},"/user/newUser":{"post":{"operationId":"UserController_newUser","summary":"Create system user.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"responses":{"201":{"description":"Created."},"400":{"description":"Bad request."},"401":{"description":"Unauthorized."}},"tags":["user"],"security":[{"token":[]}]}},"/authenticate/login":{"post":{"operationId":"AuthenticateController_login","summary":"login to system","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Login"}}}},"responses":{"201":{"description":"Authorized."},"400":{"description":"Bad request."},"401":{"description":"Unauthorized."}},"tags":["authenticate"]}}},"info":{"title":"Cloud Native Backend API","description":"The Meeting Center API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"token":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Room":{"type":"object","properties":{"name":{"type":"string"},"building":{"type":"string"},"capacity":{"type":"number"},"area":{"type":"number"},"eating":{"type":"boolean"}},"required":["name","building","capacity","area","eating"]},"User":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["Admin","User"]}},"required":["firstName","lastName","email","username","password","role"]},"Login":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}}}